Fix Typo in "with-eslint" Example docs (#28601)
Earlier it was:

```bash
npx create-next-app --example with-eslint with-eslint
# or
yarn create next-app --example with-eslint with-eslint
```

It was confusing for me, 
as i couldn't identify which one is *example-name* and which one is *folder-name* (for *with-eslint*).

I checked other example folders and was able to digest the stuff.
But if a new programmer comes who is not familiar with next/npx,
this could have took some time to understand.

And Hence i modified the cmd it to  the below, for better understanding:

```bash
npx create-next-app --example with-eslint with-eslint-app
# or
yarn create next-app --example with-eslint with-eslint-app
```

## Documentation / Examples

- [X] Make sure the linting passes