Delete duplicate SECURITY.md (#43856)
This file is redundant because it is already [inherited from the org](https://github.com/vercel/.github/blob/main/SECURITY.md). 

This will allow us to manage the security policy in one place for all repos.

See [the docs](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file) for more info.