Autolabel with linear when labeled with kind: bug (#44774)
Creates a new workflow and action that auto labels with `linear` whenever we apply `kind: bug`.
It's using JS action instead of docker, so it should be quite fast.

I tested this on my fork and it works.

Also created the action with TS so we can use it as a starting point for next actions so they can all be typed.