Improve TS experience for metadata (#45493)
This PR adds the corresponding TS plugin logic for the `metadata` export
field. In short, it provides the same IntelliSense experience to that
export value even if it isn't typed manually with `Metadata`. It is
implemented via an in-memory TypeScript LanguageServiceHost, that
attaches the `Metadata` type to the original node (if it's not typed by
the user) and handles the actual IntelliSense requests. Our plugin here
only services as a proxy.

That said, in our docs (and even in our JSDoc of metadata), we should
still recommend people to type it manually.

Also note that `generateMetadata` isn't enhanced with this yet.

## Bug

- [ ] Related issues linked using `fixes #number`
- [ ] Integration tests added
- [ ] Errors have a helpful link attached, see
[`contributing.md`](https://github.com/vercel/next.js/blob/canary/contributing.md)

## Feature

- [ ] Implements an existing feature request or RFC. Make sure the
feature request has been accepted for implementation before opening a
PR.
- [ ] Related issues linked using `fixes #number`
- [ ]
[e2e](https://github.com/vercel/next.js/blob/canary/contributing/core/testing.md#writing-tests-for-nextjs)
tests added
- [ ] Documentation added
- [ ] Telemetry added. In case of a feature if it's used or not.
- [ ] Errors have a helpful link attached, see
[`contributing.md`](https://github.com/vercel/next.js/blob/canary/contributing.md)

## Documentation / Examples

- [ ] Make sure the linting passes by running `pnpm build && pnpm lint`
- [ ] The "examples guidelines" are followed from [our contributing
doc](https://github.com/vercel/next.js/blob/canary/contributing/examples/adding-examples.md)