Return a special result when a route can't be resolved (#45668)
In Turbopack, we defer to the Next.js router for rewrites and redirects.
We need to be able to differentiate results where the Next.js router
found a corresponding route and where it didn't, in order to redirect
the latter to assets or rewrite them to 404.

Right now, with `nextConfig.basePath = "base"`, the router will return
the same result for `/base/1` and `/1`. With this diff, the router would
return `none` for `/1`, which lets us differentiate on the Turbopack
side.