chore: improve PR labeling, fix, sort, add new team member (#46675)
## What?

This PR has its main goal to improve the experience of opening/reviewing
a PR.

It improves the PR template, as well as extends PR auto-labeling.


## Why?

There are 2 parts to the PR template:

- Individual contributors: The current template accommodates for the
most common cases individual contributors use to contribute. (I.e.
fixing bugs, updating/adding examples, addressing documentation issues).
The new template keeps this at the top.
- Next.js team: As maintainers, we should try to explain PRs as if
someone with less context would review them. This should also help with
backreferencing PRs in issues and follow-up PRs and make it easier to
follow changes, understand trade-offs better at the time, and help new
team members catch up over time, etc.

## How?

The template is written in a comment instead of Markdown, so when the PR
is opened, most of the template is hidden to make it less verbose.

The checkboxes we've had before are not shown, but we already auto-label
PRs based on the path of the files that have been touched, so the
checkboxes showed redundant information.


[Slack
thread](https://vercel.slack.com/archives/C04DUD7EB1B/p1677677802109289),
Closes NEXT-741