chore: add App Router template to bug report (#46872)
### What?

Link to
https://github.com/vercel/next.js/tree/canary/examples/reproduction-template-app-dir
from our bug report template.

### Why?

To lower the barrier to adding a reproduction even when reporting an App
Router-related issue, which is currently behind a flag. (Pointed out in
#46852)

### How?

Added a link in the issue template. The remaining question is, should we
add anything else to
https://github.com/vercel/next.js/tree/canary/examples/reproduction-template-app-dir
to make creating a reproduction easier?

Closes NEXT-789
Related #46852